Add workflow to build and test the Nix shell environment
The workflow will check whether the current QMK firmware version can be
built using the Nix shell environment.  As a side effect, it will also
populate the binary cache, so that any users of this Nix shell
environment would be able to install the cached packages instead of
running a lengthy build themselves.