Use `rv32imac` and `ilp32` for the RISC-V toolchain
The `riscv32-embedded` and `riscv64-embedded` toolchains provided by
Nixpkgs do not support multilib, and an attempt to enable the multilib
support was unsuccessful (although the gcc and newlib packages seem to
be compiled with multilib support, the library search path is not set
properly, therefore ld tries to use wrong libraries and fails).  As a
workaround, compile the `riscv32-embedded` toolchain with the `arch` and
`abi` values required for GD32VF103.