Use `env.CACHIX_NAME` instead of `secrets.CACHIX_SETTINGS_JSON`
Storing the Cachix cache name in a secret (so that it could be
reconfigured without changing the actual workflow file) turned out to be
more trouble that it's worth (one of the problems is that the secrets
are not passed to workflow runs triggered by pull requests, which is
mostly fine for actual secrets, but not for the cache name which is
actually not a secret).  Use a more traditional way of passing the cache
name through the environment (the value is specified in the YAML file).