Bump `nixpkgs` to the most recent version
The `nixpkgs` version was frozen on a snapshot from March 2021; this
was useful to avoid expensive rebuilds for things like `avr-gcc`, but
also resulted in using obsolete versions of tools like `git` and
underlying libraries like `openssl` (which accumulated multiple CVEs
since that time).  Bump the `nixpkgs` snapshot to the current version
(2022-04-18), so that the packages used for build get updated.

With the new `nixpkgs` snapshot the `avr-gcc` 8.x binaries can be
obtained from the official NixOS binary cache, so the only things that
need local rebuilds (or trusting a third-party binary cache) should be
the Python dependencies.