envrc: Avoid creating multiple cache entries in `.direnv/`
When using `use nix`, current versions of `nix-direnv` include the the
version of `nixpkgs` found in `NIX_PATH` in the cache entry name under
`.direnv/`, presumably because `shell.nix` might depend on it.  However,
in this project `shell.nix` uses a locked version of Nixpkgs and does
not depend on the `nixpkgs` version available in `NIX_PATH`, therefore
the above behavior of `nix-direnv` results only in useless environment
rebuilds which give exactly the same result.  Changing `NIX_PATH` to
point `nixpkgs` to `/dev/null` works around the problem (`nix-direnv`
uses `unknown` instead of the Nixpkgs version, and therefore uses a
fixed name for the cache entry, therefore upgrading the system version
of Nixpkgs does not trigger a cache rebuild).