Replace `qmk-dotty-dict` with `dotty-dict`
The `qmk-dotty-dict` module was just a workaround for a Windows-specific
encoding issue (https://www.github.com/pawelzny/dotty_dict/issues/69),
and that issue was solved in `dotty-dict` release 1.3.1, therefore
`qmk-dotty-dict` can now be replaced with `dotty-dict`.  Actually this
replacement is required to use `qmk` 1.1.1, because that version depends
on `dotty-dict`, and having both `dotty-dict` and `qmk-dotty-dict` in
the dependency tree results in conflicts.