Sync `nix/pyproject.toml` with the `develop` branch of `qmk/qmk_firmware`
Some updates of the `requirements.txt` and `requirements-dev.txt` files
in `qmk/qmk_firmware` have been missed; resync `nix/pyproject.toml` to
the state of those files in the `develop` branch as of now.

The only difference is that `requirements.txt` still mentions the old
`qmk-dotty-dict` module, while `nix/pyproject.toml` is updated to use
`dotty-dict`; this update is required, because both of those modules are
essentially different versions of the same module, and attempting to
install both of them in the same Python environment causes conflicts.
If https://www.github.com/qmk/qmk_firmware/pull/18117 gets merged, it
would remove even this difference.