ci: Check for the `qmk setup` success using `grep`
Unfortunately, `qmk setup` does not return the exit code of the internal
`qmk doctor` invocation and just pretends that everything is successful
if if reached the `qmk doctor` step.  To check for problems during that
`qmk doctor` invocation, save the output of `qmk setup` in a file and
then look for the expected `QMK is ready to go` string there.

The `shell: bash` setting is needed to make sure that the `pipefail`
option is enabled, so the exit code of `qmk setup` also gets checked.