Fix `qmk doctor` invocation from `qmk setup`
The last step of `qmk setup` is running `qmk doctor` to check for any
problems in the resulting configuration (on non-Nix systems this step
actually handles installation of dependencies).  However, the Python
code was trying to use `sys.argv[0]` to invoke the same `qmk` script as
itself, and this did not work in Nix, because `sys.argv[0]` points to
the actual entry point script instead of the wrapper created by
`mkPoetryEnv`, therefore a proper Python module path was not set.

Explicitly invoking the interpreter from `sys.executable` fixes the
problem, because in Nix `sys.executable` is set from the Nix-specific
`NIX_PYTHONEXECUTABLE` environment variable, which is set in the `qmk`
wrapper from the environment created by `mkPoetryEnv`, and the path in
`sys.executable` points to the `python` wrapper from the same
environment, which sets `NIX_PYTHONPATH` correctly, therefore the
invoked interpreter used the proper Python module path, and `qmk doctor`
can run successfully.