Bump `nixpkgs` to the current `nixos-unstable` snapshot (2023-05-02)
The `nixpkgs` version was frozen on a snapshot from March 2021; this was
useful to avoid expensive rebuilds for things like `avr-gcc`, but also
resulted in using obsolete versions of tools like `git` and underlying
libraries like `openssl` (which accumulated multiple CVEs since that
time).  Even worse, the Python packaging ecosystem like `poetry` changed
so much that old versions of `nixpkgs` are basically unusable (many
packages need updated overrides which are included in new versions of
`poetry2nix`, but those versions of `poetry2nix` are not compatible with
old `nixpkgs`).

Bump the `nixpkgs` snapshot to the current version (2023-05-02), so that
the packages used for build get updated, and tools like `poetry2nix` can
actually work without major workarounds.

With the new `nixpkgs` snapshot the `avr-gcc` 8.x binaries for
`x86_64-linux` can be obtained from the official NixOS binary cache, so
the only things that need local rebuilds (or trusting a third-party
binary cache) should be the Python dependencies.  Unfortunately, this is
not the case for `x86_64-darwin`, because `avr-gcc` 8.x does not build
properly on Hydra due to excess number of compile warnings (the build
log size exceeds the limit).