Update poetry2nix override for `pillow` == 9.5.0
There is a bug in the `pillow` 9.5.0 release that breaks the build:

  https://github.com/nix-community/poetry2nix/issues/1139
  https://github.com/python-pillow/Pillow/pull/7069

Apply the patch with the upstream fix when building version 9.5.0.

The old override is still kept, even though the main part of that
override was included in nix-community/poetry2nix#689 (the problem is
that `buildInputs` in the poetry2nix override does not include all
packages that are in the nixpkgs package for `pillow`, and therefore
`preConfigure` might add references to nonexistent packages).