ci: Bump Nixpkgs in `nixPath` to `nixos-24.05`
NixOS 23.11 is no longer supported, and the `nixos-23.11` channel had
unfortunately been bumped in a way that made it mostly unusable on
Darwin (there are no cached binary packages for the latest revision, the
`nixpkgs-23.11-darwin` channel points to some older revision).  Update
Nixpkgs to an actually supported version to avoid that problem.