Switch back to `clang-tools` 11
QMK CI still uses `clang-format` version 11; using a different version
results in annoying mismatches that need to be fixed when submitting PRs
(otherwise the format check rejects the PR).  The `clang-tools_11`
package is no longer present in recent versions of Nixpkgs, therefore an
old Nixpkgs snapshot is used to get that package.